bugfix: Now doesn't crash when hashchecking large torrents - Downloading of large torrents needs to be tested though
svn path=/trunk/bitsharp/; revision=66617

+++ src/MonoTorrent.Client/Managers/FileManager.cs
@@ -254,7 +254,7 @@ namespace MonoTorrent.Client
         {
             lock (this.hashBuffer)
             {
-                int pieceStartIndex = this.pieceLength * pieceIndex;
+                long pieceStartIndex = (long)this.pieceLength * pieceIndex;
                 int bytesRead = this.Read(this.hashBuffer, 0, pieceStartIndex, this.pieceLength);
                 return hasher.ComputeHash(this.hashBuffer, 0, bytesRead);
             }
